[export] [telemetry] Instrument library
Add a new library to manage metrics: this is expected to subsume both the StatsPlugin architecture, and the original core stats stack.

The primary artifacts here are InstrumentDomain and its inner Storage type, and MetricsQuery.

InstrumentDomain manages a set of metrics that have a common suite of labels. Its Storage type then maintains values of those metrics for each distinct set of labels.

MetricsQuery allows querying against all domains for metrics. Importantly it also has a 'CollapseLabel' mechanism, which allows omitting a label from the output wherever it exists, and summing across all instances of Storage that contained the remaining labels. This is the mechanism we'll use to regain `optional` labels from our APIs.

Histograms are omitted from the implementation initially.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/786767276](http://cl/786767276) [cl/789008914](http://cl/789008914)

PiperOrigin-RevId: 786767276