[export] Mark Fuchsia OS as a gRPC small client for binary size optimization. Add an opt-out flag --//:exclude_small_client.
Most of the Fuchsia-powered devices I'm aware of are heavily storage-constrained, and I know some Fuchsia package using gRPC use the existing build flags to strip features. This will make it the default, similar to Android and iOS.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/789838066](http://cl/789838066)

PiperOrigin-RevId: 789838066